// src/api-client/models/branch-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BranchResponse
 */
export interface BranchResponse {
    /**
     * 
     * @type {number}
     * @memberof BranchResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'location'?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'phone'?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'status'?: BranchResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'createdBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'updatedBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof BranchResponse
     */
    'updatedAt'?: string;
}

export const BranchResponseStatusEnum = {
    Active: 'ACTIVE',
    Inactive: 'INACTIVE'
} as const;

export type BranchResponseStatusEnum = typeof BranchResponseStatusEnum[keyof typeof BranchResponseStatusEnum];